requirements.txt
acme==1.21.0
aiohttp==3.9.1
aiohttp-retry==2.8.3
aiosignal==1.3.1
alembic==1.13.1
anyio==4.3.0
apache-airflow==2.9.0
apache-airflow-providers-common-io==1.3.0
apache-airflow-providers-common-sql==1.11.1
apache-airflow-providers-fab==1.0.2
apache-airflow-providers-ftp==3.7.0
apache-airflow-providers-http==4.10.0
apache-airflow-providers-imap==3.5.0
apache-airflow-providers-smtp==1.6.1
apache-airflow-providers-sqlite==3.7.1
apispec==6.6.0
apturl==0.5.2
arabic-reshaper==3.0.0
argcomplete==3.2.3
asgiref==3.8.1
asn1crypto==1.5.1
aspose-words==23.8.0
async-timeout==4.0.3
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
Babel==2.14.0
bcrypt==3.2.0
beautifulsoup4==4.12.2
black==24.4.2
blinker==1.7.0
Brlapi==0.8.3
Brotli==1.1.0
cachelib==0.9.0
certbot==1.21.0
certifi==2022.6.15
cffi==1.15.0
channels==4.1.0
channels-redis==4.2.0
chardet==3.0.4
charset-normalizer==2.0.12
click==8.1.2
clickclick==20.10.2
colorama==0.4.5
colorlog==4.8.0
command-not-found==0.3
commonmark==0.9.1
ConfigArgParse==1.5.3
configobj==5.0.6
ConfigUpdater==3.2
connexion==2.14.2
constantly==23.10.4
coreapi==2.3.3
coreschema==0.0.4
cron-descriptor==1.4.3
croniter==2.0.3
cryptography==41.0.7
cssselect2==0.7.0
cupshelpers==1.0
daphne==4.1.2
dbus-python==1.2.18
defer==1.0.6
Deprecated==1.2.13
dill==0.3.1.1
distlib==0.3.7
distro==1.7.0
distro-info==1.1+ubuntu0.2
Django==4.0.5
django-admin-rangefilter==0.10.0
django-ckeditor==6.7.0
django-cors-headers==3.13.0
django-crispy-forms==2.0
django-crontab==0.7.1
django-environ==0.11.2
django-filter==22.1
django-js-asset==2.2.0
django-model-utils==1.1.0
django-oauth-toolkit==2.3.0
django-rest-auth==0.9.5
django-rest-swagger==2.2.0
django-sslserver==0.22
django-versatileimagefield==2.2
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
dnspython==2.6.1
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
docutils==0.16
drf-yasg==1.21.7
duplicity==0.8.21
email_validator==2.1.1
environ==1.0
exceptiongroup==1.1.3
fasteners==0.14.1
filelock==3.13.1
Flask==2.2.5
Flask-AppBuilder==4.4.1
Flask-Babel==2.0.0
Flask-Caching==2.1.0
Flask-JWT-Extended==4.6.0
Flask-Limiter==3.5.1
Flask-Login==0.6.3
Flask-Session==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.2.1
fonttools==4.47.0
freetype-py==2.3.0
frozenlist==1.4.0
fsspec==2023.12.2
future==0.18.2
google-re2==1.1
googleapis-common-protos==1.63.0
googletrans==4.0.0rc1
greenlet==3.0.3
grpcio==1.62.1
grpcio-tools==1.62.0
gunicorn==21.2.0
gyp==0.1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
html5lib==1.1
httpcore==0.9.1
httplib2==0.20.2
httpx==0.13.3
humanize==4.0.0
hyperframe==5.2.0
hyperlink==21.0.0
idna==2.10
imgkit==1.2.3
importlib-metadata==6.11.0
importlib_resources==6.4.0
incremental==22.10.0
inflection==0.5.1
iso-639==0.4.5
isort==5.13.2
itsdangerous==2.1.2
itypes==1.2.0
jeepney==0.7.1
Jinja2==3.1.2
joblib==1.3.2
josepy==1.10.0
Js2Py==0.74
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jwcrypto==1.5.5
keyboard==0.13.5
keyring==23.5.0
langdetect==1.0.9
langid==1.1.6
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lazy-object-proxy==1.10.0
limits==3.10.1
linkify-it-py==2.0.3
lockfile==0.12.2
louis==3.20.0
lxml==5.1.0
macaroonbakery==1.3.1
Mako==1.3.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.21.1
marshmallow-oneofschema==3.1.1
marshmallow-sqlalchemy==0.28.2
mdit-py-plugins==0.4.0
mdurl==0.1.2
monotonic==1.6
more-itertools==10.2.0
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.4
mypy-extensions==1.0.0
netifaces==0.11.0
networkx==3.2.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.99
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
olefile==0.46
openapi-codec==1.3.2
opentelemetry-api==1.24.0
opentelemetry-exporter-otlp==1.24.0
opentelemetry-exporter-otlp-proto-common==1.24.0
opentelemetry-exporter-otlp-proto-grpc==1.24.0
opentelemetry-exporter-otlp-proto-http==1.24.0
opentelemetry-proto==1.24.0
opentelemetry-sdk==1.24.0
opentelemetry-semantic-conventions==0.45b0
ordered-set==4.1.0
oscrypto==1.3.0
outcome==1.2.0
packaging==23.2
pandas==2.0.1
paramiko==2.9.3
parsedatetime==2.6
pathspec==0.12.1
pdfkit==1.0.0
pendulum==3.0.0
pexpect==4.8.0
Pillow==9.1.1
pipenv==2023.9.8
platformdirs==4.2.2
pluggy==1.4.0
prison==0.2.1
protobuf==4.25.3
psutil==5.9.8
psycopg2==2.9.9
ptyprocess==0.7.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycairo==1.20.1
pycparser==2.22
pycups==2.0.1
pydyf==0.11.0
PyGithub==1.55
Pygments==2.17.2
PyGObject==3.42.1
pyHanko==0.20.1
pyhanko-certvalidator==0.24.1
PyICU==2.8.1
pyjsparser==2.7.1
PyJWT==2.8.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyodbc==4.0.32
pyOpenSSL==24.1.0
pyparsing==2.4.7
pypdf==4.3.1
pyphen==0.15.0
pypng==0.20220715.0
pyRFC3339==1.1
pyrsistent==0.18.1
PySocks==1.7.1
python-apt==2.4.0+ubuntu3
python-bidi==0.5.1
python-daemon==3.0.1
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
python-decouple==3.6
python-dotenv==1.0.1
python-gitlab==3.3.0
python-magic==0.4.27
python-nvd3==0.15.0
python-slugify==8.0.4
pytz==2024.1
pyxdg==0.27
PyYAML==6.0.1
qrcode==7.4.2
redis==5.0.7
referencing==0.31.1
reportlab==4.2.2
requests==2.31.0
requests-toolbelt==0.9.1
reviews==0.4.2
rfc3339-validator==0.1.4
rfc3986==1.5.0
rich==12.2.0
rich-argparse==1.4.0
rlPyCairo==0.3.0
rpds-py==0.18.0
scikit-learn==1.4.1.post1
scipy==1.12.0
SecretStorage==3.3.1
selenium==4.9.0
sentry-sdk==1.23.1
service-identity==24.1.0
setproctitle==1.3.3
simple-colors==0.1.5
simplejson==3.19.2
six==1.16.0
slack-sdk==3.26.1
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
SQLAlchemy==1.4.52
SQLAlchemy-JSONField==1.0.2
SQLAlchemy-Utils==0.41.2
sqlparse==0.4.2
ssh-import-id==5.11
svglib==1.5.1
sympy==1.12
systemd-python==234
tabulate==0.9.0
tenacity==8.2.3
termcolor==2.4.0
text-unidecode==1.3
texttable==1.6.4
threadpoolctl==3.3.0
time-machine==2.14.1
tinycss2==1.2.1
tomli==2.0.1
torch==2.2.1
trio==0.22.2
trio-websocket==0.10.3
triton==2.2.0
twilio==8.11.0
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.9.0
tzdata==2022.1
tzlocal==5.2
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
uc-micro-py==1.0.3
ufw==0.36.1
unattended-upgrades==0.1
unicodecsv==0.14.1
universal_pathlib==0.2.2
uritemplate==4.1.1
uritools==4.0.2
urllib3==1.26.18
usb-creator==0.3.7
uvicorn==0.15.0
virtualenv==20.24.5
virtualenv-clone==0.3.0
wadllib==1.3.6
weasyprint==60.2
webencodings==0.5.1
websocket-client==1.2.3
Werkzeug==2.2.3
wkhtmltopdf==0.2
wrapt==1.14.1
wsproto==1.2.0
WTForms==3.1.2
xdg==5
xhtml2pdf==0.2.13
xkit==0.0.0
yarl==1.9.4
zipp==3.18.1
zope.component==4.3.0
zope.event==4.4
zope.hookable==5.1.0
zope.interface==6.4.post2
zopfli==0.2.3
